pyproject.toml
[project]
name = "departic"
version = "0.2.1"
description = "Departure-driven EV charge planning via iCal and EVCC."
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
keywords = ["ev", "charging", "evcc", "ical", "calendar", "soc", "trip-planning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "pydantic>=2.0",
    "apscheduler>=3.10",
    "requests>=2.32",
    "icalendar>=6.0",
    "python-dateutil>=2.9",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/nickybulthuis/departic"
Issues = "https://github.com/nickybulthuis/departic/issues"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.7",
    "responses>=0.25",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/departic"]

[tool.pytest.ini_options]
addopts = "--cov=departic --cov-report html --cov-fail-under=95 --cov-branch --log-cli-level=INFO"
log_cli = true
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
omit = [
    "src/departic/__main__.py",
    "src/departic/main.py",
    "src/departic/conftest.py",
    "src/departic/*_test.py",
    "src/departic/web/*_test.py",
]

[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "@overload",
    "raise NotImplementedError",
    '\.\.\.',
]

[tool.ruff]
line-length = 88
target-version = "py313"
extend-exclude = ["LICENSE"]

[tool.ruff.lint]
select = [
    'A',    # flake8-builtins
    'ASYNC',# flake8-async
    'B',    # flake8-bugbear
    'BLE',  # flake8-blind-except
    'C4',   # flake8-comprehensions
    'DTZ',  # flake8-datetimez
    'E',    # pycodestyle errors
    'ERA',  # eradicate (no commented-out code)
    'EXE',  # flake8-executable
    'F',    # pyflakes
    'FLY',  # flynt (f-string conversion)
    'FURB', # refurb (modernisation hints)
    'G',    # flake8-logging-format
    'I',    # isort
    'ICN',  # flake8-import-conventions
    'LOG',  # flake8-logging
    'PERF', # perflint
    'PIE',  # flake8-pie
    'PLC',  # Pylint conventions
    'PLE',  # Pylint errors
    'PLW',  # Pylint warnings
    'PT',   # flake8-pytest-style
    'PTH',  # flake8-use-pathlib
    'RET',  # flake8-return
    'RUF',  # Ruff-specific rules
    'S',    # flake8-bandit
    'SIM',  # flake8-simplify
    'T',    # flake8-debugger / flake8-print
    'TC',   # flake8-type-checking (move TYPE_CHECKING-only imports)
    'TRY',  # tryceratops
    'UP',   # pyupgrade
    'W',    # pycodestyle warnings
    #'ARG',  # unused arguments — common and intentional in many patterns
    #'D',    # pydocstyle — too verbose for this project
    #'N',    # pep8-naming — too opinionated
]

ignore = []

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"*_test.py" = ["S101"]         # assert is fine in tests
"src/departic/models.py" = ["TC"]  # Pydantic needs annotations imported at runtime
